added product tests, moved logig to base page from main
